src/python/requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
aiounittest==1.3.0
annotated-types==0.7.0
anyio==4.4.0
attrs==23.2.0
azure-core==1.30.1
azure-cosmos==3.2.0
azure-identity==1.17.1
azure-search==1.0.0b2
azure-storage-blob==12.7.0
azure-storage-queue==12.1.5
blinker==1.8.2
botbuilder-azure==4.15.0
botbuilder-core==4.15.0
botbuilder-integration-aiohttp==4.15.0
botbuilder-schema==4.15.0
botframework-connector==4.15.0
botframework-streaming==4.15.0
certifi==2024.6.2
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==42.0.7
defusedxml==0.7.1
distro==1.9.0
Flask==3.0.3
frozenlist==1.4.1
gunicorn==22.0.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
isodate==0.6.1
itsdangerous==2.2.0
Jinja2==3.1.4
jsonpickle==1.4.2
jsonschema==4.23.0
jsonschema-path==0.3.3
jsonschema-specifications==2023.12.1
lazy-object-proxy==1.10.0
MarkupSafe==2.1.5
more-itertools==10.3.0
msal==1.30.0
msal-extensions==1.2.0
msrest==0.7.1
multidict==6.0.5
nest-asyncio==1.6.0
numpy==2.0.1
oauthlib==3.2.2
openai==1.30.5
openapi-core==0.19.2
openapi-schema-validator==0.6.2
openapi-spec-validator==0.7.1
packaging==24.0
parse==1.20.2
pathable==0.4.3
portalocker==2.10.1
prance==23.6.21.0
pybars4==0.9.13
pycparser==2.22
pydantic==2.7.3
pydantic-settings==2.3.4
pydantic_core==2.18.4
PyJWT==2.8.0
PyMeta3==0.5.1
python-dotenv==1.0.1
PyYAML==6.0.1
referencing==0.35.1
requests==2.32.3
requests-oauthlib==2.0.0
rfc3339-validator==0.1.4
rpds-py==0.19.1
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.8
ruff==0.5.2
semantic-kernel==1.3.0
six==1.16.0
sniffio==1.3.1
tqdm==4.66.4
typing_extensions==4.12.1
urllib3==1.26.18
Werkzeug==3.0.3
yarl==1.9.4
